One more stage

+++ hqm-ranked-database/DbModels/Setting.cs
@@ -29,7 +29,7 @@ namespace hqm_ranked_backend.Models.DbModels
         public int RequiredGamesCount { get; set; } = 0;
         public string WebUrl { get; set; } = String.Empty;
         public string TelegramBotToken { get; set; } = String.Empty;
-        public int TelegramGroupId { get; set; } = 0;
+        public long TelegramGroupId { get; set; } = 0;
         public int NotificationThreadId { get; set; } = 0;
         public int NewsThreadId { get; set; } = 0;
         public string TelegramJoinLink { get; set; }
